object_tracing.c: fix argument type
* ext/objspace/object_tracing.c (make_unique_str): fix argument type.
  use long for string length.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@40975 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/objspace/object_tracing.c
@@ -37,7 +37,7 @@ struct allocation_info {
 };
 
 static const char *
-make_unique_str(st_table *tbl, const char *str, int len)
+make_unique_str(st_table *tbl, const char *str, long len)
 {
     if (!str) {
 	return NULL;
